+++ frontend/src/components/Navigation/Navigation.jsx
@@ -11,38 +11,22 @@ import SignupFormModal from '../SignupFormModal/SignupFormModal';
 
 export default function Navigation( {isLoaded} ) {
   const dispatch = useDispatch();
-const user = useSelector(state => state.session.user)
+const user = useSelector(state => state.session?.user)
 
 
 
-const sessionLinks = user ? (
-  <>
-    <li>
-      <ProfileButton user={user} />
-    </li>
-  </>
-) : (
-  <>
-     <li>
-        <OpenModalButton
-          buttonText="Log In"
-          modalComponent={<LoginFormModal />}
-        />
-      </li>
-      <li>
-        <OpenModalButton
-          buttonText="Sign Up"
-          modalComponent={<SignupFormModal />}
-        />
-      </li>
-  </>
-);
+
   return (
     <ul>
     <li>
       <NavLink to="/">Home</NavLink>
     </li>
-    {isLoaded && sessionLinks}
+
+    {isLoaded && (
+        <li>
+          <ProfileButton />
+        </li>
+      )}
   </ul>
   )
 }
